restructuring of files, further polishing
the github page now includes the Achievements Plus HUD, so i put in the two different folders to match the main releases of 3AIR with Extra Mushrooms. Sawnic now works with the mod regardless of mod order thanks to a quick fix related to the Origins SSZ cutscene fix. Corrected an issue involving infinite uses of the Lightning Shield's Double Jump with dropdashing, and allows Super to use it properly if the option is set in the AIR default settings. Adds options to the mod menu and increases the mod version to 1.6.0 for a new release.