level select improving (read desc. for more info)
added an option in the mod that allows the user to force the base scripts of the Level Select (this makes it so you don't need to re-prioritize the mods that end up breaking stuff, and prevents me from needing to add more of that junk in the future). edited the letter "X" in the custom font to fit in more. fixed a potential error in which the game would attempt to draw a background that doesn't exist ("origins_s&k_bg"). made it so the S&K level select background is always displayed behind other custom backgrounds in the case that screen widths are wider than the level select background (the S&K background was updated to accommodate for this). removed a bit of useless code that was commented out. refined the conditions for playing level select music. added 2 more level select music options that enable music played in Blue Spheres. (one uses a track from Sound Additions, which requires you to have the mod on, the other changes the music depending on if you have the continue music set to S3 or S&K). changed the positioning of the left half of the level select text to make it perfectly centered when using a game screen width of 320 pixels. added in two hex values next to the character continue icons so people can make sure they are actually switching out characters when pressing C. (this MAY be useful when Extra Slot support happens later down the line)