LBZ category, changeable Big Arms hit count
someone requested the ability to change the amount of hits Big Arms takes so i thought it'd be cool to do that. has support with Bossbar as long as this mod is above that. putting LBZ stuff into its own category prevents the misc. category from being extremely bloated, and it also has enough stuff to stand on its own, i'd think.